requirements.txt
azure-search-documents
fastapi
uvicorn
sqlalchemy
pandas
psycopg2-binary
python-dotenv
python-multipart
pymupdf
langchain
langchain_openai